added UI for custom link
